Add string Add utilies for hash

+++ Source/BIEngine/StdLib/StdLib.h
@@ -2,6 +2,6 @@
 
 namespace BIEngine {
 
-using SizeT = unsigned int;
+using SizeT = unsigned long long;
 
 }
